Add ExeptionMiddleware

+++ TestProject.Logic/Services/Document/DocumentService.cs
@@ -1,5 +1,8 @@
 ﻿using AutoMapper;
+using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Logging;
+using System.Text;
+using TestProject.Core.Infrastructure.Exceptions;
 using TestProject.Logic.Services.Document.Models;
 using TestProject.Repositories.Entities;
 using TestProject.Repositories.Interfaces;
@@ -9,33 +12,69 @@ namespace TestProject.Logic.Services.Document;
 public class DocumentService : IDocumentService
 {
     private readonly IMapper _mapper;
-    private readonly ILogger<DocumentService> _logger;
+   // private readonly ILogger<DocumentService> _logger;
     private readonly IDbDocumentRepository _documentRepository;
 
-    public DocumentService(IDbDocumentRepository documentRepository, IMapper mapper, ILogger<DocumentService> logger)
+    //public DocumentService(IDbDocumentRepository documentRepository, IMapper mapper, ILogger<DocumentService> logger)
+    //{
+    //    _mapper = mapper;
+    //    _logger = logger;
+    //    _documentRepository = documentRepository;
+    //}
+
+    public DocumentService(IDbDocumentRepository documentRepository, IMapper mapper)
     {
         _mapper = mapper;
-        _logger = logger;
         _documentRepository = documentRepository;
-
     }
 
-    public Task<List<ulong>> Insert(List<DocumentModel> documents)
+    public async Task<List<ulong>> InsertAsync(IFormFile file)
+    {
+        string textFromFile = await ReadFromFileAsync(file);
+        if (!string.IsNullOrEmpty(textFromFile))
+        {
+            List<DocumentModel> documents = ConvertToListDocumentModels(textFromFile);
+
+            return await Task.FromResult(SaveToCollection(documents));
+        }
+        else
+        {
+            return null; //ToDo: можно ли возвращать null и как его тогда обрабатывать выше?
+        }
+    }
+
+    public List<ulong> SaveToCollection(List<DocumentModel> documents)
     {
         List<ulong> ids = new();
 
-        foreach(DocumentModel documentModel in documents)
+        foreach (DocumentModel documentModel in documents)
         {
             var id = _documentRepository.Add(_mapper.Map<DbDocument>(documentModel));
             ids.Add(id);
         }
 
-        return Task.FromResult(ids);
+        return ids;
     }
 
-    public async Task<List<DocumentModel>?> GetByContractNumber(string contractNumber)
+    public async Task<DocumentModel> GetById(ulong documentId)
     {
-        var documents = _documentRepository.GetByContractNumber(contractNumber);
+        DbDocument document = _documentRepository.GetById(documentId);
+        if(document == null)
+            throw new NotFoundException(typeof(DocumentModel), documentId);
+
+        return await Task.FromResult(_mapper.Map<DocumentModel>(document));
+    }
+
+    public async Task<List<DocumentModel>> GetByContractNumberAsync(string contractNumber)
+    {
+        if (contractNumber == null)
+        {
+            throw new InvalidArgumentException("ContractNumber is required.");
+        }
+
+        List<DbDocument> documents = _documentRepository.GetByContractNumber(contractNumber);
+        if (documents.Count == 0)
+            throw new NotFoundException(typeof(DocumentModel), contractNumber);
 
         List<DocumentModel> result = new();
 
@@ -44,17 +83,18 @@ public class DocumentService : IDocumentService
             result.Add(_mapper.Map<DocumentModel>(document));
         }
 
-        return await Task.FromResult(result); //ToDo: прокинуть асинхронность везде
+        return await Task.FromResult(result);
     }
 
-    public async Task<List<DocumentModel>?> GetByUserId(string userId)
+    public async Task<List<DocumentModel>> GetByUserIdAsync(string userId)
     {
         if(userId == null)
         {
-            _logger.LogInformation(""); //ToDo: дописать лог
-            return null;
+            throw new InvalidArgumentException("UserId is required.");
         }
-        var documents = _documentRepository.GetByUserId(userId);
+        List<DbDocument> documents = _documentRepository.GetByUserId(userId);
+        if (documents.Count == 0)
+            throw new NotFoundException(typeof(DocumentModel), userId);
 
         List<DocumentModel> result = new();
 
@@ -66,8 +106,59 @@ public class DocumentService : IDocumentService
         return await Task.FromResult(result);
     }
 
-    public Task<ulong> DeleteByUserIdOrContractNumber(string? userId, string? contractNumber)
+    public Task<ulong> DeleteByUserIdOrContractNumberAsync(string userId, string contractNumber)
     {
+        if (string.IsNullOrEmpty(userId) && string.IsNullOrEmpty(contractNumber))
+            throw new InvalidArgumentException("One of the parameters must not equal an empty string and null");
+
         return Task.FromResult(_documentRepository.DeleteByUserIdOrContractNumber(userId, contractNumber));
     }
+
+    private List<DocumentModel> ConvertToListDocumentModels(string text)
+    {
+        List<DocumentModel> documents = new();
+        var array = text.Split("\r");
+        foreach (var item in array)
+        {
+            DocumentModel model = new();
+            var tmpArray = item.Split("--");
+            for (int i = 0; i < tmpArray.Length; i++)
+            {
+                switch (i)
+                {
+                    case 0:
+                        model.UserId = tmpArray[0].Trim(' ').Replace("\n", "");
+                        break;
+                    case 1:
+                        model.LastName = tmpArray[1].Trim(' ').Replace("\n", "");
+                        break;
+                    case 2:
+                        model.FirstName = tmpArray[2].Trim(' ').Replace("\n", "");
+                        break;
+                    case 3:
+                        model.ContractNumber = tmpArray[3].Trim(' ').Replace("\n", "");
+                        break;
+                    case 4:
+                        model.WithdrawalAmount = Convert.ToDecimal(tmpArray[4].Trim(' ').Replace("\n", ""));
+                        break;
+                }
+            }
+            documents.Add(model);
+        }
+
+        return documents;
+    }
+
+    private async Task<string> ReadFromFileAsync(IFormFile file)
+    {
+        long length = file.Length;
+
+        using var fileStream = file.OpenReadStream();
+        byte[] buffer = new byte[length];
+        await fileStream.ReadAsync(buffer, 0, (int)file.Length);
+
+        return Encoding.Default.GetString(buffer);
+    }
+
+    
 }
